src/pyproject.toml
[tool.poetry]
name = "linebot-subsidy"
version = "0.1.0"
description = ""
authors = ["FionnKuo"]
readme = "README.md"
packages = [{include = "linebot_subsidy"}]

[tool.poetry.dependencies]
python = "^3.11"
pymysql = "^1.1.0"


[tool.poetry.group.dev.dependencies]
aiohttp = "3.8.4"
aiosignal = "1.3.1"
anyio = "3.7.0"
async-timeout = "4.0.2"
attrs = "23.1.0"
certifi = "2023.5.7"
charset-normalizer = "3.1.0"
click = "8.1.3"
colorama = "0.4.6"
fastapi = "0.97.0"
frozenlist = "1.3.3"
future = "0.18.3"
gunicorn = "20.1.0"
h11 = "0.14.0"
idna = "3.4"
line-bot-sdk = "2.4.2"
multidict = "6.0.4"
pydantic = "1.10.9"
python-dotenv = "1.0.0"
requests = "2.28.2"
sniffio = "1.3.0"
starlette = "0.27.0"
typing-extensions = "4.6.3"
urllib3 = "1.26.16"
uvicorn = "0.22.0"
yarl = "1.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
